+++ src/main/java/com/t1membership/item/dto/searchOneItem/SearchOneItemRes.java
@@ -5,6 +5,8 @@ import com.t1membership.item.constant.ItemSellStatus;
 import com.t1membership.item.domain.ItemEntity;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -17,7 +19,7 @@ public class SearchOneItemRes {
 
     private String itemName;
 
-    private int itemPrice;
+    private BigDecimal itemPrice;
 
     private int itemStock;
